refactor: split credit/cli.py into cli/ package; address DJ PR comments
- Split 2652-line credit/cli.py into credit/cli/ package:
  _common.py  — PBS defaults, prompts, logging, agent bash tools
  _convert.py — credit convert/init handlers (imports at top)
  _submit.py  — all PBS/submission handlers (imports at top, env vars removed)
  _ask.py     — AI assistant agent/ask handlers
  _plot.py    — credit plot/metrics handlers
  _parser.py  — argument parser and main() entry point (init first)
  __init__.py — re-exports all public symbols for backward compat

- Rewrite scripts/build_env_casper.sh as parameterized general script
- Remove scripts/build_env_derecho.sh and scripts/casper_gen2.sh
- Fix monkeypatch targets in test_cli.py to use correct module namespaces
